services/trading-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clashx-trading-backend"
version = "0.1.0"
description = "ClashX trading backend powered by Pacifica SDK"
requires-python = ">=3.11,<3.13"
dependencies = [
  "alembic>=1.14.1",
  "base58>=2.1.1",
  "cryptography>=44.0.2",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.10.0",
  "sqlalchemy>=2.0.36",
  "python-dotenv>=1.0.1",
  "psycopg[binary]>=3.2.6",
  "httpx>=0.28.1",
  "privy-client>=0.5.0",
  "solders>=0.19.0",
  "websockets>=14.1",
  "pacifica-sdk>=0.1.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.9.10",
  "mypy>=1.15.0",
  "pytest>=8.3.5"
]

[tool.pdm.scripts]
lint = "ruff check src"
format = "ruff format src"
typecheck = "mypy src"
migrate = "alembic upgrade head"
makemigration = "alembic revision --autogenerate"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff.lint]
ignore = ["B008", "UP017"]

[tool.pytest.ini_options]
testpaths = ["tests"]
